+++ src/laravel/HasNeighbors.php
@@ -3,6 +3,7 @@
 namespace Pgvector\Laravel;
 
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\MissingAttributeException;
 
 trait HasNeighbors
 {
@@ -38,8 +39,7 @@ trait HasNeighbors
     {
         $id = $this->getKey();
         if (!array_key_exists($column, $this->attributes)) {
-            // TODO use MissingAttributeException when Laravel 9 no longer supported
-            throw new \OutOfBoundsException('Missing attribute');
+            throw new MissingAttributeException($this, $column);
         }
         $value = $this->getAttributeValue($column);
         return static::whereKeyNot($id)->nearestNeighbors($column, $value, $distance);